Variabilize more the keystore settings addition
Given it's convenient to not use `--stdin` with the ES keystore to fix #1349,
this commit makes the keystore settings addition different for ES than
for Kibana and the APM Server.

Only the Elasticsearch keystore supports settings additions using a file
path. Kibana and APM Server supports only to pass the value through
stdin and use the --stdin flag.