Update stack version in docs and samples with a bash script (#1413) (#1415)
Add a bash script to automatically update all referenced stack versions in docs and samples.
The script was applied with 7.2.0 (7.2.1 should be available soon, but isn't yet).

I had to deal with sed tricks to make it compatible with both OSX sed and GNU sed.
To avoid dealing with curl responses from Elasticsearch (GET /), I chose to shortcut the version section of the response to a simple {...}.

Another way of doing this would be to templatize docs and samples entirely, and generate them from the template. Which would be a bit cleaner, but also many more files. I thought this first step would maybe be good enough (until we have too many special cases).


* Add a bash script to replace all stack versions in docs and samples

* Run the script with 7.2.0

* Remove the version section of ES API response

* Set a version independant custom image tag example

* Replace image by tag