Use storage class with WaitForFirstConsumer for all E2E tests (#1414)
The default storage class with GKE does not have WaitForFirstConsumer
set by default. This means PVs get bound before a pod is scheduled in
its zone, which could lead to all nodes of a cluster being scheduled on
the same host. This happens when all the PVs have been randomly created
with the same node affinity (same zone => same host).
As a quick fix, we will use a custom default storage class with
WaitForFirstConsumer for all the E2E tests.