Orchestrate zen1 and zen2 settings for StatefulSets (#1262)
Change the way we setup zen1 and zen2 settings to adapt to
StatefulSets upscale and downscale.
Also improve StatefulSet downscale code, which is moved to its own file,
and add a few helper functions around pods and master nodes.

In a nutshell:

set minimum_master_nodes (zen1) and initial_master_nodes (zen2) in the configuration of new nodes to create, corresponding to the expected nodes that will be created. The setting is automagically updated in existing nodes configuration , where it is ignored until nodes restart.
update minimum_master_nodes (zen1) and voting_config_exclusions (zen2) whenever we remove a node.
For now, we don't (yet) do any change on rolling upgrades, which are limited to a single node at a time at the moment. This only impacts sset scale up/down.