Only update zen2 settings if there is no v6 master node
When doing any zen2 API call (voting config exclusions), we checked if
there is at least one v7+ master in the cluster in order to proceed.
This is wrong because if the current master is a v6 node, the call will
fail.
In a cluster with mixed v6 -> v7 versions, the current master is
supposed to be a v6 node. Zen2 APIs can be called once there is no more
v6 node in the cluster, at which point the current master is a v7 node.

For more context, see
https://github.com/elastic/cloud-on-k8s/issues/1281#issuecomment-520491832