Handle zen1 minimum_master_nodes special case for 2->1 masters (#1551)
* Handle zen1 minimum_master_nodes special case for 2->1 masters

Instead of always updating zen1 before removing a master node, which can
lead to split brain scenarios, make sure we update zen1 if we are moving
from 2 to 1 master nodes.

In other scenarios (eg. 3->2), we want to update minimum_master_nodes
after the node is removed. This works fine as long as we remove one
master at a time.

We check the actual running master pods in the cluster to determine if
we are downscaling a master when 2 masters are currently running. This
is not a perfect solution since we are relying on potentially
out-of-date k8s cached information, but we cannot do something 100%
realtime perfect.

* Remove duplicate zen1 compatibility check

* Emit an event on zen1 2->1 masters situation

* Specify the unsafe operation nature in the event

* Refactor test StatefulSet creation

* Ignore readiness of the pods when upgrading zen1 mmn