Add tests and comments for lazy ESState implementation
Add some comments on all functions, and unit test ESState behaviour, by
relying on a mocked ES client.

I'm pretty sure this ESState thing will be improved and maybe handled
differently in the future, but in the meantime it's best to have it unit
tested correctly.

I hesitated to switch to a proper Mock tool for the ES Client, but I'm
actually quite happy with the current "manual mock" approach.