Add an E2E test for ES v6 to v7 rolling upgrade (#1625)
* Add MutationOptions to expect or not a rolling update

When a rolling update happens, a node goes down without triggering data
migration first. It is expected that the cluster will become red if some
shards have 0 replicas.
Let's specify it when running the test. If we expect rolling upgrades to
happen, we should create shards with 1 replica. Else, if we expect data
migration to happen, we should create shards with 0 replicas.

* Test ES version 6.8.0 to 7.2.0 rolling upgrade