Create new pods with latest revision (#1620)
* Upscale cluster with newest revision

When we upscale a cluster from eg. 3 to 5 nodes, and change the spec at
the same time, the 2 new nodes were created with the old spec.
It makes sense to create new nodes with the new spec directly.

In order to do so, we also update the rolling update partition to the
value of the StatefulSet replicas (before the upscale).
In the above example, the partition will be updated to 3: the 2 new
nodes will be created with the newest revision, while the previous 3
nodes will stay at their current revision, until handled with the
rolling upgrade.

In case a rolling upgrade was already in progress (ie. partition equals
1), we still update it to 3. The rolling upgrade is temporarily "paused"
but will go through again.

* Fix typo