Compute number of replicas in E2E tests based on mutation topology
Automatically compute the data integrity index number of replicas, based
on the cluster topology (initial and mutated).
If single-node cluster, use 0 replicas (otherwise the cluster cannot be
green).
If a rolling upgrade should happen, use 1 replica (for the cluster to
stay green during the mutation).
Otherwise, use 0 replicas, to ensure the cluster goes red if a data
migration does not happen.