Make sure we update sset template hash label before updating
We had a bug where the StatefulSet template hash that we use for
StatefulSet comparison purpose wasn't updated during downscales and
upgrades, where we change replicas and partitions.

This is not a problem with the current codebase, but starts becoming one
once we upscale only one master node at a time.

Let's make sure we update it.