Rolling upgrade master nodes last
During an upgrade for eg. v6 to v7, a cluster with a single master node
will form a v7 cluster that v6 data nodes won't be able to join.

As specified in ES doc:
> Similarly, if you run a testing/development environment with only one
master node, the master node should be upgraded last. Restarting a
single master node forces the cluster to be reformed. The new cluster
will initially only have the upgraded master node and will thus reject
the older nodes when they re-join the cluster. Nodes that have already
been upgraded will successfully re-join the upgraded master

An "easy" way to fix that is to upgrade wannabe master nodes last.