Re-bootstrap single master 7.x clusters when upgrading from 6.x
When upgrading single master clusters from 6.x to 7.x we are effectively reconstructing the cluster.
The new 7.x node is not joining a new cluster as it would be the case for a rolling upgrade on an HA cluster,
but it rather needs to be bootstrapped again as if it were a new 7.x cluster.

We control bootstrap configuration via the cluster uuid annotation on the Elasticsearch CRD, this commit resets
that annotation on single master upgrades to 7.x thereby regenerating bootstrap configuration.