Set NSS_SDB_USE_CACHE=no to avoid memory growth
On affected systems[1], the readinessProbe costs 5mb of memory per invocation which is observed in some cases[2] to never be reclaimed. This is a bug in the kernel and/or kubernetes, but in the meantime, this env var can help prevent the readinessProbe from causing memory usage growth and eventually causing Kubernetes to evict pods.

[1] Certain versions of CentOS and GKE are known affected.
[2] It is observed in production that for pods without a strict memory `limit`, the container's file system cache (dentry) is not reclaimed appropriately, causing the kubelet to believe the pod is using far more memory than is actually being used by processes in the pod.

Fix unit tests to sort env vars so that comparisons will succeed.

References: https://github.com/elastic/cloud-on-k8s/issues/1635