Update ES resource phase to operational
The ES orchestration phase is set to operational when the ES custom
resource is reconciled, that is to say there is no more changes to be
applied by the operator.
Given the ES resource is a set of statefulsets, it's reconciled if all
actual statefulsets and all their associated pods have the expected
statefulset template hash label.
The template hash includes the number of replicas so it reflects when
there is an upscale or a downscale. The check at the pod level reveals
if there is a rolling upgrade in progress.