When using OnDelete strategy don't rely on current/updated revision
When using on delete current revision is never reset to update revision just looking that the update/current revision therefore does not work when reverting to previous revision and gives constant false positives after an initial update. Only updated replicas != replicas expresses the fact that an update is still pending.