When using OnDelete strategy don't rely on current/updated revision (#1732)
When using an OnDelete strategy in ssets current revision is never reset to update revision. Just looking that the revision therefore does not work when reverting to a previous revision and gives constant false positives after an initial update. Only updated replicas != replicas expresses the fact that an update is still pending.