Always enable file-based user auth (#1698)
The operator relies on file realm for creating all the internal users
including the built-in elastic user. This commit makes sure the file
realm is always enabled in case a user wants to configure additional
realms.
The order of the file realm is set to -100 to favor the fact that it is
used first but there is no complete guarantee about that. The user can
still configure a realm with an order less than -100 but it is very
unlikely given this is not suggested in the documentation (order starts to 0).
The main known issue could be to configure a realm that relies on a
system (e.g. Active Directory) that has a temporary lock-out period
after several successive failed login attempts, as being checked first
(https://www.elastic.co/guide/en/elastic-stack-overview/current/trouble-shoot-active-directory.html).

This change supports the two syntaxes for the realm settings depending the major version (https://www.elastic.co/guide/en/elasticsearch/reference/current/breaking-changes-7.0.html#include-realm-type-in-setting).