Remove PVCs when removing nodes (#1736)
Let's not keep PVCs around when Elasticsearch nodes are removed, since
it can be very confusing if there are reused in a different context (eg.
new cluster with the same old name, or upscale long after a downscale).

This is done in 2 ways:
- set the ownerRef of Claims in the StatefulSet resource to the ES resource, to delete PVCs on cluster deletion.
- delete unused PVCs, based on their naming conventions and what we
expect to exist from actual & expected StatefulSets.

Note we do not remove PVs, only PVCs. PVs deletion is left to the
storage class reclaim policy responsibility.