Bind ES Pod readiness to a node-local endpoint (#1777)
This changes the ES endpoint we request in the readiness probe to
'/_cat/nodes?local' instead of '/'.

The HTTP request is therefore restricted to that node only, and does not
go over another node. This should prevent situations where all v6.X Pods
go in a non-ready status temporarily when the master node gets killed.