Set a dynamic MinAvailable value for the default PDB (#1775)
Reconcile the default PDB to have its MinAvailable value consider the current number of Pods we expect to run, according to the StatefulSet specs.

This is done right after handling nodes upscale: we work with the returned (updated) StatefulSets. Summing their replicas give us the maximum number of nodes we should have running. Some of them may not be running yet, but that's OK since it will lead to a higher MinAvailable value (-> no disruption allowed).

By default, we want to set MinAvailable to NumberOfPods - 1, so only one Pod can be disrupted.
However, in cases such as:

- cluster not green (best-effort based on out-of-date information)
- only 1 master node
- only 1 data node
- we don't want any Pod to be disrupted.

Side-effects of this PR in the code:

- return the updated StatefulSets at the end of the upscale phase
- add a typed ElasticsearchUnknownHealth const


* Split pdb reconciliation into multiple smaller functions

* Don't override spec if user-provided

* Extract buildPDBSpec to its own function

* Set PDB MinAvailable from ssets replicas

* Delete then recreate if PDB spec changes

* Prevent disruption if the cluster is not healthy

* Return updated ssets after upscale, to use for PDB computation

* Add ElasticsearchUnknownHealth as a type

* Unit test the PDB reconciliation

* Fix upscale unit test for returned ssets

* Prevent disruption of the single data node

* Set ownerReference on the PDB resource for automatic deletion

* Make linter happy

* Fix typos

* Rename function for consistency

* Prevent disruption of the single ingest node

* Switch to hash-based comparison

* Fix typo