Wait for the rolling upgrade to be over after keystore e2e check
When running the kesytore E2E tests, we chain several rolling upgrades,
in between which we check keystore entries on each node. It turns out
this test is a bit flaky since we sometimes fail, waiting for more than
5 minutes.

Since we don't wait for the rolling upgrade to be fully over before
going into the next secure settings change test, the 2nds test needs to
account for the 1st test rolling upgrade to be done. Which obviously
makes it longer to run.

Let's wait for rolling upgrades to be over before moving on with the
next keystore check.

Should help with (maybe fix?)
https://github.com/elastic/cloud-on-k8s/issues/1325.