Use more than 1 node on rolling upgrade E2E tests
Doing a rolling upgrade of a 1-node cluster gives a short period of time
where the cluster becomes red, while shards are initializing after the
upgrade.

This is something we could consider in the cluster health check, but I
think it makes more sense to do the cluster RAM increase/decrease tests
with a 3-nodes cluster, to ensure the cluster is not disrupted during
the rolling upgrade.