Don't monitor health if single node rolling upgrade
Disable continuous health check monitoring if we're doing a rolling
upgrade of a 1-node cluster.

This test was flappy, since a 1-node cluster upgrade will temporarily be
Red while shards are initializing after the upgrade.

The PR also introduces an optional "Skip" field in the TestStep. If
defined, a function returning True will disable the associated test.