Restore e2e test failures on cluster health retrieval
We disabled error reporting (leading to e2e test failure) when we cannot
retrieve the health.
A "valid case" to not being able to retrieve the cluster health is when
we're restarting/removing the master node of a v6 cluster. During leader
election, the cluster is temporarily unavailable to such requests.
This is way better with v7 clusters and zen2, for which unavailability
is made a lot smaller.

Let's restore that check, but make sure we ignore any errors resulting
from a v6 cluster upgrade.