Allow all pods to be upgraded if none is ready
Bypass any rolling upgrade check and budget if some pods need to be
upgraded, and none of the existing pod is ready.

This allows moving on with clusters where all pods as misconfigured.

Side-effect: it's possible to quickly change the spec of a cluster while
it starting up for the first time, or if it's completely down.

Since the check is based on Pod readiness, there is a race condition
where the readiness we get does not match the actual one on the
apiserver. In which case we should normally not do the forced rolling
upgrade. This edge case is ignored here.