Force upgrade all Pods if bootlooping or pending (#1827)
* Allow all pods to be upgraded if none is ready

Bypass any rolling upgrade check and budget if some pods need to be
upgraded, and none of the existing pod is ready.

This allows moving on with clusters where all pods as misconfigured.

Side-effect: it's possible to quickly change the spec of a cluster while
it starting up for the first time, or if it's completely down.

Since the check is based on Pod readiness, there is a race condition
where the readiness we get does not match the actual one on the
apiserver. In which case we should normally not do the forced rolling
upgrade. This edge case is ignored here.

* Add e2e tests for forced upgrades

* Make linter happy

* Refine force upgrade heuristic for Pending and bootlooping Pods

Instead of a general "all Pods are not ready" check, this commit refines
the logic to force-upgrade Pods if:
* they are all Pending (implies not Ready)
* they are all not Ready with a restartCount >= 1

This should minimize chances to accidentally upgrade a non-ready Pod
that should not.

* Make linter happy

* Don't move on with standard rolling upgrade if ES cannot be reached

Otherwise, we get errors because we cannot fetch shards allocation
settings.