Force upgrade pods of a same StatefulSet
Prior to that change, we trigger a force upgrade for Pods that are
either Pending or bootlooping, if all Pods in the cluster are also
Pending or bootlooping.
A situation where Pods of the first StatefulSet are fine but Pods of the
2nd StatefulSet are bootlooping can still happen.
This fixes it by checking the state of all Pods of the same StatefulSet,
instead of considering all Pods in the cluster.