Reorder E2E tests checks (#1926)
Run CheckESNodesTopology and CheckESVersion before CheckESHealthGreen 
to avoid a race condition between these checks and
the operator when running mutation tests.
The race condition can cause theses checks to be run against an old
(unmutated) cluster which is not intended. The topology and the version
can be validated whether expected pods are actually there.
Get rid of CheckESReachable, which is now implied by CheckESNodesTopology.