Check expected Pods in E2E tests
Add a function that looks at all StatefulSets and Pods in the k8s
cluster, and returns an error if:
* some pods are missing
* there are extra pods around
* pods are not ready yet
* pods that should have been upgraded are not upgraded (yet)

If some pods are not ready yet (eg. still Pending because they cannot
mount a PersistentVolume), we output a pretty-printed pod Status in the
error message, for easier human parsing.

This check runs before any other mutation check, so we ensure other
checks happen **after** the cluster has been mutated.

To check if the mutation correctly happened, the approach in this commit
is to:
* annotate pods before mutation, wich a hash of the corresponding ES
Version + NodeSet specification
* make sure the hash of all existing ready Pods either matches the
current builder spec (no mutation should happen), or is empty (pod was
upgraded hence recreated with no annotation)

This does not catch upgrades due to keystore changes, but this test
has its own logic.