Refactor expectations with proper garbage collection
This commit is composed of:
- a bugfix where expectations were kept in-memory forever
- a refactoring of the expectations code

1. Garbage collection

This commit fixes the following bugs in the previous implementation:

* generation expectations were kept in memory forever for all
StatefulSets (including deleted ones or ones with no upgrade events)
* generation expectations were kept in memory forever for deleted
clusters
* deletion expectations may be kept in memory forever for deleted
clusters if the deletion happens right after a pod deletion

Generation expectations are now cleared as soon as satisfied.
Expectations for a given cluster are garbage-collected when that cluster
gets removed (reconciliation following a Delete event).

2. Refactoring

* Split deletions from generations in their own file
* Create a thread-safe per-cluster expectations map, so individual
expectations struct don't need to care about their own cluster
* Simplify deletion expectations logic
* Add a big comment block to explain why we use these expectations
* Make generation expectations work with what we have in the cache
instead of working with what it takes as argument. Similar to how pod
deletions expectations already worked. Because of that, we need to make
sure in the code that we fetch StatefulSets **after** checking
expectations. Otherwise we would check expectations on StatefulSets in
the cache, but work with a different list of StatefulSets (eg. updated
ones after upscale, not reflected yet in the cache).