Use the official Elasticsearch Go client.
This replaces the implementations of the `elasticsearch/client.Client` interfaces for v6 and v7 with a single implementation that spans v6, v7 and v8 APIs using the official Go client (https://github.com/elastic/go-elasticsearch).

Since the official client is still a low-level client, we still keep the `client/model.go` to deal with marshalling/unmarshalling.