Check resource version when deleting a Pod during force-upgrades
This commit adds an additional precondition when we delete a Pod during
a force-upgrade: we require the Pod version in the apiserver to be the
same as the cached Pod version we want to delete.

This helps avoiding situations were we would, for example, delete a
Pending pod that is actually not Pending anymore.