Perform StatefulSets deletions before replicas downscale (#2062)
* Perform StatefulSets deletions before replicas downscale

We had a bug where a StatefulSet with 0 replicas could be kept forever
under particular conditions (eg. cannot be removed since there is a risk
to remove the last master node... even though there is 0 replicas).

This commit fixes it by decoupling StatefulSets deletions from
StatefulSet downscales.
A StatefulSet deletion is only performed when the actual StatefulSet has
0 replicas. Our expectations mechanism ensures there is no Pod still
alive for that StatefulSet (eg. if it was just downscaled from 1 to 0
replicas). So that StatefulSet can be safely garbage-collected.

StatefulSets deletions are now ran before handling downscales,
regardless of any downscale invariant constraint.
The downscale code is a bit simplified since it does not need to handle
StatefulSets deletions anymore.

* Add a unit test for sset already deleted

* Make linter happy