Remove v1alpha1 CRD version & generate a single trivial all-in-one flavor (#2119)
We decided to get rid of the v1alpha1 version of the CRDs for the next release.

This PR removes the `v1alpha1` codepath, remove any kustomize patches for `v1alpha1`, and regenerates the CRDs to contain `v1beta1` only.

Since we now have only one version (and we plan the next v1 schema to be the same as the current v1beta1), there's no need for multiple CRD flavors. We can generate a single favor with the trivialVersions option, using validation for the v1beta1 schema.

I introduced a `patches` folder containing all kustomize stuff to:
- merge all CRDs into one
- remove the podTemplate properties, in order to:
    - make the `kubectl apply` annotation smaller
    - be resilient to any pod spec changes in various k8s versions
    - avoid the problem with `nil` containers in the podTemplate spec

A file `patches/all-in-one.yaml` is automatically generated, and includes the merged patched version of all CRDs. To be reused, for example in `make run` or by the deployer.
It is also used by integration tests, which I modified slightly to retrieve and parse the content of `patches/all-in-one.yaml` automatically.

The Makefile is adapted accordingly.