Add v1alpha1 in the list of versions in CRDs
Removing v1alpha1 from this list prevents users to successfully upgrade
from alpha -> beta -> GA version of the operator. This is because in
their situation, the v1alpha1 verison is listed in status.storedVersion.
We cannot easily patch the status only when using `kubectl apply` as
upgrade mechanism, and patching the entire CRD fails the validation
webhook that checks versions in the status.

This commit patches the output CRD to list the v1alpha1 version, but
mark it with `served: false` so it cannot be used anymore.