Set volumeClaimTemplate namespace to match ES namespace
The latest version of `controllerutil.SetControllerReference` does a
safety check on owner vs. object namespace mismatches.
We technically don't need to include a namespace in the
VolumeClaimtemplates of a StatefulSet, but since we know in which
namespace the PVC will be created anyway, it's safe to include the
namespace so we pass the function safety check.