Use a bugfixed version of controllerutil.SetControllerReference (#2211)
* Fix controllerutil.SetControllerReference to ignore apiVersions

Temporarily duplicate the fixed version of
`controllerutil.SetControllerReference` in our own code, until we get a
proper release of controller-runtime with the fix.

At which point we can safely remove this file and rely on
`controllerutil.SetControllerReference` again.

controller-runtime PR including the fix:
https://github.com/kubernetes-sigs/controller-runtime/pull/707

* Set volumeClaimTemplate namespace to match ES namespace

The latest version of `controllerutil.SetControllerReference` does a
safety check on owner vs. object namespace mismatches.
We technically don't need to include a namespace in the
VolumeClaimtemplates of a StatefulSet, but since we know in which
namespace the PVC will be created anyway, it's safe to include the
namespace so we pass the function safety check.