Inherit existing sset.volumeClaimTemplates ownerReferences
When we update the Elasticsearch CRD to a new version, existing
ownerReferences in StatefulSets.volumeClaimTemplates still point to the
"old" Elasticsearch apiVersion (eg. `v1beta1` instead of `v1`).

Since the volumeClaimtemplate section is immutable, the ownerRef
apiVersion cannot be updated so the StatefulSet update is rejected.
In that situation, it is fine to just keep the existing ownerRef
targeting the deprecated Elasticsearch apiVersion. It does not prevent
resource garbage collection to happen as expected.

This commit also changes the way we recently set a Namespace to the
claim, to not error out in `SetControllerReference`. We don't need that
namespace (and it breaks backward compatibility), so we now just set it
temporarily for `SetControllerReference` to not return any error.