Workaround subresources status update error
A bug in K8s < 1.15 prevents status subresource updates to be successful
if the resource to update has a different storedVersion in etcd.

This commit works around the bug by catching the error (HTTP 422 -
resource invalid), and attempt an update of the entire resource instead
of the status sub-resource only.