(not ready to merge) Create CRDs V1 (#2184)
Bump the current v1beta1 CRD to a separate v1 CRD with the exact same content. As such, it's fine to rely on the OpenAPI validation schema of the v1 version to validate the v1beta1 resources.

I tried to make sure every single reference we had to the v1beta1 API is updated to use the v1 API now. Along with that change I tried to homogenize the imports so we always import esv1, kbv1, etc. instead of using a mix of estype, esv1beta1, elasticsearchv1beta1, v1beta1, etc.

* Add v1 API codebase

* Unset storage version for v1beta1 resources

* Fix Elasticsearch webhook path

* Regenerate CRDs with v1 version

* Regenerate webhook manifests with v1 webhook

* Use v1 APIs in the entire codebase & homogenize imports

* Adapt golangci exceptions to v1 apis

* Use v1 CRD in samples

* Regenerate API docs to cover v1 crds

* Default to using crd v1 in the documentation

* Fix typo: corev1

* Regenerate api docs with the fixed version with sorting