Increase rolling upgrade unavailabilty threshold
We used to have a 60sec threshold where we consider it's OK for a
cluster to be unavailable when the master node gets killed.
This test is very flaky, especially on 6.8.x and 7.1.x clusters where
the unavailability is usually around 50sec under good conditions.

I bumped the value to 120sec: this is really an arbitrary value, mostly
to avoid flakiness of the test while keeping the current test.

https://github.com/elastic/cloud-on-k8s/issues/2262 is still open so we
figure out a better way to deal with it.