Retry mutation update in E2E tests
Updating the Elasticsearch resource may return an error if a conflict
happens (the operator updating the resource concurrently).

Let's retry if that happens, by calling the usual `test.Eventually()`.

Technically it means we also retry on error on the `Get()`, but I think
it's fine and keeps the code simple.