Don't upgrade pods if empty StatefulSet UpdateRevision
We had a bug were some Pods of a newly created StatefulSet would be
force upgraded because they are Pending, and we consider they need to be
upgraded because their revision does not match the StatefulSet update
revision.
That's because the StatefulSet updateRevision is empty on initial
StatefulSet creation.

This commit fixes it by not considering empty StatefulSet
updateRevisions when looking for Pods to upgrade.