Reuse the same upscaleState across StatefulSets
We had a bug where the second StatefulSet containing master nodes would
have its replicas set to 0, because it is not allowed to add additional
masters.

Turns out that's because we passed a new upscaleState (by value), whose
innitialization to allow masters creation only happens once. The first
StatefulSet is allowed master nodes creation thanks to the upscale
initialization, but the second StatefulSet isn't because its
upscaleState is zeroed-out and not re-initialized.

This commit fixes it by passing around a StatefulSet pointer, and adds a
unit test that fails without the fix.