Wait 30sec in the preStop hook to account for kube-proxy refresh (#2360)
* Wait 30sec in the preStop hook to account for kube-proxy refresh

Kube-proxy defaults to a 30sec iptables-sync-period (overridden to 10sec
on GKE).

For that reason, when an Elasticsearch Pod is gracefully shutdown, some
clients in the k8s cluster might still target the Pod IP, because the
local kube-proxy iptables (or ipvs) rules are not updated yet.

Let's increase our initial 1 sec additional wait time to 30 sec so we
account for those changes. It has the side-effect of slowing down
Pods deletion (hence rolling upgrades) by 30sec per Pod, but should help
keeping Elasticsearch available during rolling upgrades.

This commit also increases the Elasticsearch container gracePeriod to
account for the increased wait time. I changed it from 120 to 180
seconds (technically could have been 150, but 180 rounds up to 3
minutes).

* Fix references in the doc

* Use the default 30sec ADDITIONAL_WAIT_SECONDS in E2E test

* Add a prefix to the preStop env variables

This makes the variable names more explicit when overridden by the user
in the ES spec.

* Improve comments and docs with kube-proxy details

* Wording: dont guarantee no downtime, but minimise unavailability