Perform a synchronous request to retrieve ES cluster UUID
Since we only perform that request when:
- the cluster is not marked as bootstrapped yet
- es is supposed to be reachable

It's fine to do it synchronously, and not rely on state observed every
10 seconds asynchronously.

Doing a sync requests reduces a bit the time window where we could wrongly
assume the cluster has not been bootstrapped yet, and removes some
complexity in the code.

This commit also refactors a bit the unit tests with a single function
that should cover all cases.