Mark the Version field as required in all CRDs
Currently if the Version field is not set, we reject the resource in the
validation webhook.
If the webhook is disabled, the operator returns an error early in the
reconciliation.

This could be considered a breaking API change, but the Version field
was already required implicitly, just not caught by the OpenAPI
validation.