Customize CRDs to work around trivialVersions gen bug
A bug in controller-tools prevent validation to appear in the generated
CRDs if trivialVersions is set and multiple versions do not share the
exact same schema.
See https://github.com/kubernetes-sigs/controller-tools/issues/349.

This commit removes the usage of `crd:trivialVersions=true`, to
"manually" (through Kustomize) apply the CRD modifications so we end up
with the trivialVersion scheme. That is: a single `validation` field
matching the OpenAPISchema of the currently stored version (v1).