Add a tool to recreate a deleted cluster from existing PersistentVolumeClaims (#2551)
Add a tool called reattach-pv into our hack directory.

It allows recreating a deleted Elasticsearch cluster while reusing existing PersistentVolumes. This is intended for catastrophic failures (or accidental cluster removal) that are not supposed to happen under normal conditions. More details in README.md.

Also add a basic test to check the script behaves as expected when targeting a GKE cluster.

And include compilation of this tool as part of CI: since it relies on "internal" operator code, we may accidentally break it if we change the operator code.