Check StatefulSets observedGeneration in expectations
This commit ensures the Elasticsearch driver is not working based on a
StatefulSet that has not been processed by the StatefulSet controller
yet. This is particularly important in rolling upgrades where we check
the value of `status.updateRevision`.

When a change is made to a StatefulSet spec, the `metadata.generation`
field gets incremented by the apiserver. This concerns the spec only,
not other metadata (eg. annotations or labels).
Since Kubernetes 1.6, the StatefulSet controller sets
`status.observedGeneration` to the value of `metadata.generation`. An
easy way to know if the StatefulSet has been processed by the
StatefulSet controller yet is to compare generation vs.
observedGeneration.